# docs/swagger.yaml
basePath: /api/v1
definitions:
  entity.Area:
    properties:
      MergerName:
        description: 组合名
        type: string
      Pinyin:
        description: 拼音
        type: string
      area_code:
        description: 行政代码
        type: integer
      city_code:
        description: 区号
        type: string
      lat:
        description: 纬度
        type: number
      level:
        description: 层级
        type: integer
      lng:
        description: 经度
        type: number
      name:
        description: 名称
        type: string
      parent_code:
        description: 父级行政代码
        type: integer
      short_name:
        description: 简称
        type: string
      zip_code:
        description: 邮政编码
        type: integer
    type: object
  state.ErrorResponse:
    properties:
      code:
        description: 错误代码
        type: integer
      message:
        description: 错误描述
        type: string
    type: object
host: 127.0.0.1:8080
info:
  contact:
    email: archer.plus@qq.com
    name: Archer
    url: https://archer.plus
  description: common project.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://archer.plus
  title: API文档
  version: "1.0"
paths:
  /areas/cities/{code}:
    get:
      consumes:
      - application/json
      description: 根据省份行政代码获取城市列表
      parameters:
      - description: 省份行政代码
        in: path
        name: code
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 城市列表
          schema:
            items:
              $ref: '#/definitions/entity.Area'
            type: array
        "400":
          description: 错误信息
          schema:
            $ref: '#/definitions/state.ErrorResponse'
      summary: 获取城市列表
      tags:
      - 行政区域
  /areas/committees/{code}:
    get:
      consumes:
      - application/json
      description: 根据街道行政代码获取绝味会列表
      parameters:
      - description: 街道行政代码
        in: path
        name: code
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 居委会列表
          schema:
            items:
              $ref: '#/definitions/entity.Area'
            type: array
        "400":
          description: 错误信息
          schema:
            $ref: '#/definitions/state.ErrorResponse'
      summary: 获取居委会列表
      tags:
      - 行政区域
  /areas/districts/{code}:
    get:
      consumes:
      - application/json
      description: 根据城市行政代码获取区县列表
      parameters:
      - description: 城市行政代码
        in: path
        name: code
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 区县列表
          schema:
            items:
              $ref: '#/definitions/entity.Area'
            type: array
        "400":
          description: 错误信息
          schema:
            $ref: '#/definitions/state.ErrorResponse'
      summary: 获取区县列表
      tags:
      - 行政区域
  /areas/provinces:
    get:
      consumes:
      - application/json
      description: 获取省份列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Area'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/state.ErrorResponse'
      summary: 获取省份列表
      tags:
      - 行政区域
  /areas/streets/{code}:
    get:
      consumes:
      - application/json
      description: 根据区县行政代码获取街道列表
      parameters:
      - description: 区县行政代码
        in: path
        name: code
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 街道列表
          schema:
            items:
              $ref: '#/definitions/entity.Area'
            type: array
        "400":
          description: 错误信息
          schema:
            $ref: '#/definitions/state.ErrorResponse'
      summary: 获取街道列表
      tags:
      - 行政区域
swagger: "2.0"